Moved errors in it's own file
