+++ packages/taquito/src/contract/utils.ts
@@ -1,0 +1,7 @@
+export function computeLength(data: string | Object) {
+  if (typeof data === 'object') {
+    return Object.keys(data).length;
+  } else {
+    return 1;
+  }
+}
